requirements.txt
python-docx>=0.8.11
pypandoc>=1.11